add loader for grouped workflows
