+++ lib/cellect/server.rb
@@ -11,6 +11,7 @@ module Cellect
     require 'cellect/server/grouped_workflow'
     require 'cellect/server/user'
     require 'cellect/server/loader'
+    require 'cellect/server/grouped_loader'
     require 'cellect/server/api'
 
     class << self
